pyproject.toml
[tool.poetry]
name = "reMember"
version = "0.0.1"
description = ""
authors = ["Fabian Damken"]
readme = "README.md"

[tool.poetry.dependencies]
numpy = "^1.26.4"
pillow = "^10.3.0"
python = "^3.12"
genanki = "^0.13.1"

[tool.poetry.group.dev.dependencies]
matplotlib = "^3.9.0"

[tool.poetry.group.tools.dependencies]
black = "^24.4.2"
isort = "^5.13.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 130

[tool.isort]
profile = "black"
line_length = 130
